feat : Post Delete 기능 구현 및 Board Delete 로직 수정

+++ frontend/simple_board/lib/model/post_request_dto.dart
@@ -7,7 +7,7 @@ part 'post_request_dto.g.dart';
 @JsonSerializable()
 class PostDeleteDTO implements RequestBase {
   @JsonKey(name: 'post_id')
-  double postId;
+  int postId;
   @override
   String password;
